rename Stream to Gallery
